requirement.txt
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
streamlit==1.8.1